Move components to a separate folder
#15
